refactor COLLECTION to keep proccess save

+++ jy_quote/deploy.sh
@@ -6,8 +6,11 @@ docker build -t registry.cn-shanghai.aliyuncs.com/jqy-common/$ImageName -f jy_te
 docker push registry.cn-shanghai.aliyuncs.com/jqy-common/$ImageName
 call_remote="ssh root@172.16.7.34"
 $call_remote "docker pull registry.cn-shanghai.aliyuncs.com/jqy-common/$ImageName"
-$call_remote "docker rm -f $ImageName" 
-$call_remote "docker run -d --env-file .env \
+# 如果容器存在，则删除它
+if docker ps -a --format '{{.Names}}' | grep -q "^$ImageName$"; then
+    $call_remote "docker rm -f $ImageName"
+fi
+$call_remote "docker run -d --env-file /home/cal_code/.env \
     -v /opt/vola:/app/vola \
     --name=$ImageName\
     registry.cn-shanghai.aliyuncs.com/jqy-common/$ImageName"
